Split X.509 CRT parsing in multiple steps
This commit restructures the parsing of X.509 CRTs in the following way:

First, it introduces a 'frame' structure `mbedtls_x509_crt_frame`, which
contains pointers to some structured fields of a CRT as well as copies of
primitive fields. For example, there's a pointer-length pair delimiting the raw
public key data in the CRT, but there's a C-uint8 to store the CRT version
(not a pointer-length pair delimiting the ASN.1 structure holding the version).

Setting up a frame from a raw CRT buffer does not require any memory outside
of the frame structure itself; it's just attaches a 'template' to the buffer
that allows to inspect the structured parts of the CRT afterwards.

Note that the frame structure does not correspond to a particular ASN.1
structure; for example, it contains pointers to delimit the three parts
of a CRT (TBS, SignatureAlgorithm, Signature), but also pointers to the
fields of the TBS, and pointers into the Extensions substructure of the TBS.

It is up to debate to remove/add fields to the frame structure, but for now,
the idea is that the frame stops where parsing either requires dynamic memory
or is context-dependent in a non-trivial way. That is, it doesn't parse
Subject/Issuer fields, nor does it parse the SubjectPublicKeyInfo structure
(we could have parsed the top-level structure of the SubjectPublicKeyInfo
structure, but currently the PK layer uses the SubjectPublicKeyInfo struct
as a whole as the input/output format, so there's no gain in parsing further
here). However, we do peek at the version value before potentially parsing
the v3Extensions structure, which is an example of context-dependent parsing.

The commit introduces an internal function `x509_crt_parse_frame()` which
sets up a frame from a raw buffer, as well as several tiny helper functions
which help setting up more complex structures (Subject, Issuer, PK) from the
frame.

These functions are then put to use to rewrite the existing parsing function
`mbedtls_x509_crt_parse_der_core()` by setting up a CRT frame from the input
buffer, residing on the stack, and afterwards copying the respective fields
to the actual `mbedtls_x509_crt` structure and performing the deeper parsing
through the various helper functions.

While this introduces some stack- and code-size overhead, it'll allow
to move towards on-demand CRT parsing without too much code-duplication.